Implement ReflectionContext::addImage with section finder
Implement a version of addImage that takes in a closure responsible
for finding the sections of interest. This allows for the registration
of metadata even in situations where the object file is not complete
(for example, when generated by the JIT).

(cherry picked from commit 0a67868a1368d22741b31e36a5af27fc9872cf21)