[DebugInfo] Emit an opaque struct for specialized generic types
When emitting full debug info for bound generic structs and enums, the
compiler will emit full information for the unsubstituted generic type and
"shallow" information for the substituted type. I initially the same
information as -g for the shallow type, this patch changes that to emit
just an opaque struct type, as that has all the information we need to
process the debug information from LLDB's side.